preprocess: split into __init__, utils, and integrate
